`CoverageTestRunState` can now be shared between several class instances (e.g. for Suites).
This can be useful in case when a combined coverage report for multiple process definitions is required. Shared state is stored outside of Rule.

`CoverageTestRunState` is now an interface and created by a factory. `DefaultCoverageTestRunState` is the same as the old `CoverageTestRunState`. `AggregatedCoverageTestRunState` stores current `CoverageTestRunState`, changes it for each new test class and persists `ClassCoverage`s in `AggregatedClassCoverage`.

Report creation is slightly refactored to accept `AggregatedClassCoverage` and potentially store aggregated report in non-default directory.